Work on mistakes

+++ src/test/java/ru/netology/stats/StatsServiceTest.java
@@ -3,17 +3,16 @@ package ru.netology.stats;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
 
 class StatsServiceTest {
 
     @Test
     public void Sum() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
-        int actual = service.sum(sales);
-        int expected = 8 + 15 + 13 + 15 + 17 + 20 + 19 + 20 + 7 + 14 + 14 + 18;
+        long actual = service.sum(sales);
+        long expected = 8 + 15 + 13 + 15 + 17 + 20 + 19 + 20 + 7 + 14 + 14 + 18;
 
         Assertions.assertEquals(expected, actual);
     }
@@ -21,10 +20,10 @@ class StatsServiceTest {
     @Test
     public void Average() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
-        int actual = service.average(sales);
-        int expected = (8 + 15 + 13 + 15 + 17 + 20 + 19 + 20 + 7 + 14 + 14 + 18) / 12;
+        long actual = service.average(sales);
+        long expected = (8 + 15 + 13 + 15 + 17 + 20 + 19 + 20 + 7 + 14 + 14 + 18) / 12;
 
         Assertions.assertEquals(expected, actual);
     }
@@ -32,10 +31,10 @@ class StatsServiceTest {
     @Test
     public void monthMaxSale() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
-        int actual = service.monthMaxSale(sales);
-        int expected = 8;
+        long actual = service.monthMaxSale(sales);
+        long expected = 8;
 
         Assertions.assertEquals(expected, actual);
     }
@@ -43,10 +42,10 @@ class StatsServiceTest {
     @Test
     public void monthMinSale() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
-        int actual = service.monthMinSale(sales);
-        int expected = 9;
+        long actual = service.monthMinSale(sales);
+        long expected = 9;
 
         Assertions.assertEquals(expected, actual);
 
@@ -55,10 +54,10 @@ class StatsServiceTest {
     @Test
     public void mBelAverage() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
-        int actual = service.mBelAverage(sales);
-        int expected = 5;
+        long actual = service.mBelAverage(sales);
+        long expected = 5;
 
         Assertions.assertEquals(expected, actual);
 
@@ -67,10 +66,10 @@ class StatsServiceTest {
     @Test
     public void mAboveAverage() {
         StatsService service = new StatsService();
-        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
+        long[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
 
-        int actual = service.mAboveAverage(sales);
-        int expected = 5;
+        long actual = service.mAboveAverage(sales);
+        long expected = 5;
 
         Assertions.assertEquals(expected, actual);
     }
